requirements.txt
aiohttp>=3.9.0
aiodocker>=0.21.0
anthropic>=0.40.0
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=24.1.0
